Rename handler* modules
